Add files via upload
Evolution of polish notation as an application of stack using C. This project mainly deals with infix to postfix expression with the use of stack. 

Polish prefix notation or simply prefix notation, is a mathematical notation in which operators precede their operands. Polish notation is used as a syntax for mathematical expressions by programming language interpreters, it is readily parsed into abstract syntax trees and can, in fact, define a one-to-one representation for the same. Modern Stack-organized computers are better suited for postfix and prefix notations than their infix counterpart. The following report describes the use of Polish Notation, types, algorithms; the Infix Notation, the Prefix Notation, and the Postfix Notation. Furthermore, a simulation setup to convert Infix Notation
to Prefix or Polish Notation using stack is shown.

Introduction
 
Problem Definition

The development of polish notation is an example of how stack can be used in computer programming. The problem with Polish notation is that it is not easy to read or write. In addition, it is difficult to evolve programs that use it, because there is no standard way to do so.

Motivation of the project

The motivation for this project is to investigate the history and evolution of Polish notation, as well as to explore how this notation can be applied to solve problems using a stack. Additionally, this project will aim to understand why Polish notation has been generally abandoned in favor of other notations.
Basic description of the project

In mathematics and computer science, Polish notation (also known as normal Polish notation or Polish postfix notation) is a way of writing arithmetic expressions in which the operands (numbers) come before the operators. It is named after the Polish logician Jan Łukasiewicz, who invented Polish notation in the 1920s. Polish notation is also known as prefix notation because the operator is always placed before the operands. This is the opposite of infix notation, which is the more common way of writing arithmetic expressions, in which the operator is placed between the operands, as in 3 + 4. In Polish notation, the operator is always placed after the operands. For example, the expression 3 + 4 would be written as + 3 4 in Polish notation.

Polish notation is often used in computer science because it can be used to create parse trees, which are used to represent the structure of a programming language. In general, Polish notation is not as easy to read as infix notation, but it has some advantages. One advantage is that it eliminates the need for parentheses, which can be helpful in complicated expressions. Another advantage is that it can be easier to parse expressions written in Polish Notation.

Related Theories and Algorithms

Fundamental theories underlying the work

There are a few different ways to think about the evolution of Polish notation. One way is to consider how it has been used in different applications over time. Another way is to think about the different theoretical approaches that have been used to try to explain its origins. One thing that is clear is that Polish notation has been around for a long time. Since then, it has been used in a variety of different fields, including mathematics, computer science, and linguistics. There are a few different theories about the origins of Polish notation. One theory is that it was developed as a way to make formulas easier to read. Another theory is that it was developed as a way to make formulas easier to write. Whatever the origins of Polish notation, it is clear that it has been a useful tool in a variety of different fields. It is likely that it will continue to be used in new and innovative ways in the future.

Fundamental algorithms

Prefix/postfix notation is especially popular for its innate ability to express the intended order of operations without the need for parentheses and other precedence rules, as are usually employed with infix notation. Instead, the notation uniquely indicates which operator to evaluate first. The operators are assumed to have a fixed arity each, and all necessary operands are assumed to be explicitly given. A valid prefix expression always starts with an operator and ends with an operand.

Evaluation can either proceed from left to right, or in the opposite direction. Starting at the left, the input string, consisting of tokens denoting operators or operands, is pushed token for token on a stack, until the top entries of the stack contain the number of operands that fits to the top most operator (immediately beneath). This group of tokens at the stack top (the last stacked operator and the according number of operands) is replaced by the result of executing the operator on these/this operand(s). Then the processing of the input continues in this manner. The rightmost operand in a valid prefix expression thus empties the stack, except for the result of evaluating the whole expression. When starting at the right, the pushing of tokens is performed similarly, just the evaluation is triggered by an operator, finding the appropriate number of operands that fits its already at the stack top. Now the leftmost token of a valid prefix expression must be an operator, fitting to the number of operands in the stack, which again yields the result. As can be seen from the description, a push-down store with no capability of arbitrary stack inspection suffices to implement this parsing.

Proposed model/algorithm

Proposed model

Infix Notation

We write expression in infix notation, e.g. a - b + c, where operators are used in-between operands. It is easy for us humans to read, write, and speak in infix notation but the same does not go well with computing devices. An algorithm to process infix notation could be difficult and costly in terms of time and space consumption.

Prefix Notation

In this notation, operator is prefixed to operands, i.e. operator is written ahead of operands. For example, +ab. This is equivalent to its infix notation a + b. Prefix notation is also known as Polish Notation.

Proposed algorithms

1. Scan all the symbols one by one from left to right in the given Infix Expression.
2. If the reading symbol is operand, then immediately append it to the Postfix Expression .
3. If the reading symbol is left parenthesis ‘( ‘, then Push it onto the Stack.
4. If the reading symbol is right parenthesis ‘)’, then Pop all the contents of the stack until the respective left parenthesis is popped and append each popped symbol to Postfix Expression.
5. If the reading symbol is operator (+ , – , * , /), then Push it onto the Stack. However, first pop the operators which are already on the stack that have higher or equal precedence than the current operator and append them to the postfix . If open parenthesis is there on top of the stack then push the operator into the stack.
6. If the input is over, pop all the remaining symbols from the stack and append them to the postfix .

Team Members: Manvi Bishnu, Rohit Kumar, Moumita Maji, Sneha Dey
College: Asansol Engineering College